fix some lint errors
