+++ src/server/pkg/sync/pipe_windows.go
@@ -8,5 +8,5 @@ import (
 )
 
 func (p *Puller) makePipe(path string, f func(io.Writer) error) error {
-	return fmt.Errorf("Lazy file sync through pipes is not supported on Windows")
+	return fmt.Errorf("lazy file sync through pipes is not supported on Windows")
 }
